refactor: create actions module
